requirements.txt
pandas 
pydantic 
openai 
python-dotenv 
fastapi 
uvicorn
gunicorn
numpy
tqdm
rich
openenv